Make clippy; move `target_specific` to `concurrency`;
