+++ ractor/src/concurrency/mod.rs
@@ -84,3 +84,24 @@ pub use self::async_std_primitives::*;
 pub mod tokio_with_wasm_primitives;
 #[cfg(all(target_arch = "wasm32", target_os = "unknown"))]
 pub use self::tokio_with_wasm_primitives::*;
+
+#[cfg(all(target_arch = "wasm32", target_os = "unknown"))]
+mod target_specific {
+    /// A wrapper for [std::marker::Send] on non-`wasm32-unknown-unknown` targets, or an empty trait on `wasm32-unknown-unknown` targets.
+    /// Introduced for compatibility between wasm32 and other targets
+    pub trait MaybeSend {}
+    impl<T> MaybeSend for T {}
+    pub(crate) use web_time::SystemTime;
+}
+#[cfg(not(all(target_arch = "wasm32", target_os = "unknown")))]
+mod target_specific {
+    /// A wrapper for [std::marker::Send] on non-`wasm32-unknown-unknown` targets, or an empty trait on `wasm32-unknown-unknown` targets.
+    /// Introduced for compatibility between wasm32 and other targets
+    pub trait MaybeSend: Send {}
+    impl<T> MaybeSend for T where T: Send {}
+    pub(crate) use std::time::SystemTime;
+}
+
+#[cfg(not(feature = "async-trait"))]
+pub use target_specific::MaybeSend;
+pub(crate) use target_specific::SystemTime;
