+++ ractor/src/factory/worker.rs
@@ -14,9 +14,9 @@ use std::sync::Arc;
 use bon::Builder;
 use tracing::Instrument;
 
-use crate::concurrency::{Duration, Instant, JoinHandle};
 #[cfg(not(feature = "async-trait"))]
-use crate::MaybeSend;
+use crate::concurrency::MaybeSend;
+use crate::concurrency::{Duration, Instant, JoinHandle};
 use crate::{Actor, ActorId, ActorProcessingErr};
 use crate::{ActorCell, ActorRef, Message, MessagingErr, SupervisionEvent};
