+++ ractor/src/lib.rs
@@ -258,24 +258,3 @@ pub async fn spawn_named<T: Actor + Default>(
 ) -> Result<(ActorRef<T::Msg>, JoinHandle<()>), SpawnErr> {
     T::spawn(Some(name), T::default(), args).await
 }
-
-#[cfg(all(target_arch = "wasm32", target_os = "unknown"))]
-mod target_specific {
-    /// A wrapper for [std::marker::Send] on non-`wasm32-unknown-unknown` targets, or an empty trait on `wasm32-unknown-unknown` targets.
-    /// Introduced for compatibility between wasm32 and other targets
-    pub trait MaybeSend {}
-    impl<T> MaybeSend for T {}
-    pub(crate) use web_time::SystemTime;
-}
-#[cfg(not(all(target_arch = "wasm32", target_os = "unknown")))]
-mod target_specific {
-    /// A wrapper for [std::marker::Send] on non-`wasm32-unknown-unknown` targets, or an empty trait on `wasm32-unknown-unknown` targets.
-    /// Introduced for compatibility between wasm32 and other targets
-    pub trait MaybeSend: Send {}
-    impl<T> MaybeSend for T where T: Send {}
-    pub(crate) use std::time::SystemTime;
-}
-
-#[cfg(not(feature = "async-trait"))]
-pub use target_specific::MaybeSend;
-pub(crate) use target_specific::SystemTime;
